Update wikiscraper.py
Some wikipedia articles store the first <b> tag of the <p> tag in an <i> tag, so I implemented an if statement to check for the <i> tags contents being the same as the <b> tag. In short checks to see if the title is the same as the text if so grab the parent of the text and continue. Also some articles cause issues when trying to write the text to a text file, so I implemented a try function to check if there is a UnicodeEncodeError if so encode the text to utf-8.